requirements.txt
Flask
python-pptx
langchain
langchain-community
langchain-openai
python-dotenv
pytz